// src/api/types/settings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AppSettings = { git_path: string | null, recent_repos: Array<string>, theme: string, font_size: number, font_family: string, code_font_family: string, pull_mode: string, external_editor: string | null, enable_fsmonitor: boolean, enable_untracked_cache: boolean, };